Synchronize -> Serialize
- migrate all usage to Serialized instead of Synchronized
- remove implementations of SerializedObserver that lost the competition (performance and testing in production)
